fix(intromarkers): store silence attempt file IDs as bigint
media_files.id is bigint, so an int4 media_file_id would reject attempts
for files past the int32 range and leave them in the backfill loop. The
table migration now creates the column as bigint, and the follow-up
migration widens it on a database that already applied the integer
version.

+++ migrations/sql/20260923000630_add_intro_silence_refinement_attempts.sql
@@ -5,7 +5,7 @@
 -- settings hash: forever after a clean no-improvement result, and until
 -- retry_after after a failure.
 CREATE TABLE intro_silence_refinement_attempts (
-    media_file_id integer PRIMARY KEY REFERENCES media_files(id) ON DELETE CASCADE,
+    media_file_id bigint PRIMARY KEY REFERENCES media_files(id) ON DELETE CASCADE,
     config_hash text NOT NULL,
     file_hash text NOT NULL,
     file_size bigint NOT NULL,
